Move server tests into server dir
