+++ server/test/fixtures/hospitalStatusUpdates.json
