+++ server/test/integration/api/ambulances.js
@@ -3,7 +3,7 @@ const HttpStatus = require('http-status-codes');
 const session = require('supertest-session');
 
 const helper = require('../../helper');
-const app = require('../../../server/app');
+const app = require('../../../app');
 
 describe('/api/ambulances', () => {
   let testSession;
