+++ server/test/integration/webhooks/sffd.js
@@ -5,8 +5,8 @@ const path = require('path');
 const session = require('supertest-session');
 
 const helper = require('../../helper');
-const app = require('../../../server/app');
-const models = require('../../../server/models');
+const app = require('../../../app');
+const models = require('../../../models');
 
 describe('/webhooks/sffd', () => {
   let testSession;
