+++ server/test/unit/models/patient.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
 const helper = require('../../helper');
-const models = require('../../../server/models');
+const models = require('../../../models');
 
 describe('models.Patient', () => {
   beforeEach(async () => {
